functions/requirements.txt
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.9.0
colorama==0.4.6
cryptography==42.0.5
deprecation==2.1.0
firebase-admin==6.5.0
firebase-functions==0.4.1
flask==2.1.2
functions-framework==3.5.0
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-cloud-vision==3.7.2
google-crc32c==1.5.0
google-events==0.12.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
imageio==2.34.0
itsdangerous==2.2.0
Jinja2==3.1.3
lazy_loader==0.4
MarkupSafe==2.1.5
msgpack==1.0.8
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scikit-image==0.23.2
scipy==1.13.0
tifffile==2024.4.18
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.2
